requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
python-magic-bin ; sys_platform == "win32"
python-magic ; sys_platform != "win32"
